- RTP code changes and debit card limit api integration(UI mapping left)

+++ lib/feature/payment/request_from_new_recipient/request_from_new_recipient_view_model.dart
@@ -209,9 +209,8 @@ class RequestFromNewRecipientViewModel extends BasePageViewModel {
             purpose: purposeController.text,
             purposeDetail: purposeDetailController.text,
             amount: double.parse(ProviderScope.containerOf(context)
-                    .read(requestMoneyViewModelProvider)
-                    .currentPinValue)
-                .toInt(),
+                .read(requestMoneyViewModelProvider)
+                .currentPinValue),
             limit: limit,
             dbtrBic: dbtrBic ?? "",
             dbtrAcct: dbtrAcct ?? "",
